DjangoDrf/requirements.txt
certifi==2018.10.15
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
diff-match-patch==20121119
Django==2.0.2
django-cors-headers==2.4.0
django-crispy-forms==1.7.2
django-filter==1.1.0
django-formtools==2.1
django-guardian==1.4.9
django-import-export==1.1.0
django-reversion==3.0.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
future==0.16.0
httplib2==0.11.3
idna==2.7
itypes==1.1.0
jdcal==1.4
Jinja2==2.10
lxml==4.2.5
MarkupSafe==1.0
mysqlclient==1.3.13
odfpy==1.3.6
openpyxl==2.5.9
Pillow==5.3.0
pycryptodome==3.7.2
PyJWT==1.6.4
pytz==2018.5
PyYAML==3.13
requests==2.20.0
simplejson==3.16.0
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.24
xlrd==1.1.0
XlsxWriter==1.1.1
xlwt==1.3.0
xmltodict==0.11.0
